+++ Client/Unix/HookUnix.cpp
@@ -15,8 +15,7 @@ CHook* CHook::GetHook(QObject *parent)
 }
 
 CHookUnix::CHookUnix(QObject *parent)
-    : CHook(parent),
-    m_pNativeEventFilter(new CHookUnixNativeEventFilter())
+    : CHook(parent)
 {
     //TODO:
     //RegisterKeyboard();
@@ -29,29 +28,10 @@ CHookUnix::~CHookUnix()
 
 int CHookUnix::RegisterKeyboard()
 {
-    if(!m_pNativeEventFilter)
-    {
-        Q_ASSERT(false);
-        return -1;
-    }
-#if QT_VERSION >= QT_VERSION_CHECK(5, 0, 0)
-    qApp->installNativeEventFilter(m_pNativeEventFilter);
-#else
-    if(QCoreApplication::eventDispatcher())
-        QCoreApplication::eventDispatcher()->installNativeEventFilter(m_pNativeEventFilter);
-#endif
     return 0;
 }
 
 int CHookUnix::UnRegisterKeyboard()
 {
-    if(!m_pNativeEventFilter) return -1;
-
-#if QT_VERSION >= QT_VERSION_CHECK(5, 0, 0)
-    qApp->removeNativeEventFilter(m_pNativeEventFilter);
-#else
-    if(QCoreApplication::eventDispatcher())
-        QCoreApplication::eventDispatcher()->removeNativeEventFilter(m_pNativeEventFilter);
-#endif
     return 0;
 }